Moved semantic checks inside appropriate models
This will help with multistart issues in
MLFreeNBHMM, but not in MLConstrainedNBHMM
